requirements.txt
pandas >= 0.24.2
openpyxl >= 2.5.12
numpy >= 1.16.3
geopy >= 1.14.0
urllib3 >= 1.21.1
jinja2 >= 2.10.1
pymysql >= 0.9.3
redis >= 3.2.1
SQLAlchemy >= 1.3.3
requests >= 2.19.1
cryptography >= 2.6.1
python-dateutil >= 2.6
xlrd >= 1.2.0
pathlib >= 1.0.1
shapely >= 1.6.4.post2
flake8 >= 3.7.7
mysql-connector-python >= 8.0.16
matplotlib >= 3.0.3
boto3 >= 1.9.137
bs4 >= 0.0.1
pypinyin >= 0.36.0
pyarrow >= 0.16.0

pytest >= 4.4.1
pytest-cov >= 2.7.1
pytest-pep8 >= 1.0.6
pytest-flakes >= 4.0.0
